+++ mustache_lambda.hpp
@@ -9,6 +9,7 @@ class Lambda : public mustache::Lambda {
   protected:
     virtual int getUserFunctionParamCount() = 0;
     virtual int invokeUserFunction(zval *retval_ptr, int param_count, zval params[]) = 0;
+    std::string invokeUserFunctionAsString(int param_count, zval params[]);
   public:
     std::string invoke() override;
     std::string invoke(